+++ db/migrate/20220514145133_add_middle_name_to_resident.rb
@@ -1,0 +1,5 @@
+class AddMiddleNameToResident < ActiveRecord::Migration[7.0]
+  def change
+    add_column :residents, :middle_name, :string
+  end
+end