+++ CHANGELOG.md
@@ -11,6 +11,7 @@ In place of release version numbers, we organize via deploys to Production (by D
 
 - New task: "import:abi_members"
   - Imports Lots, Properties, Resident1 & 2
+  - Add/Import Middle Name to Resident
 - Deps:
   - Add amazing_print (formatting and colors - "abc".red)